update structure project
