+++ src/components/Hero/modals/BtnRegistration.jsx
