+++ src/components/Hero/modals/ModalLastStep.staled.js
@@ -8,12 +8,12 @@ import {
   grayStroke,
   grayText,
   white,
-} from '../../utils/variables.styled';
+} from '../../../utils/variables.styled';
 import { Field, Form } from 'formik';
 import { Link } from 'react-router-dom';
-import checkbox0 from '../../assets/home/modal/checkbox0.svg';
-import checkbox1 from '../../assets/home/modal/checkbox1.svg';
-import { device } from '../../styles/device';
+import checkbox0 from '../../../assets/home/modal/checkbox0.svg';
+import checkbox1 from '../../../assets/home/modal/checkbox1.svg';
+import { device } from '../../../styles/device';
 
 export const Modal = styled.div`
     position: relative;
