+++ src/components/modalElements/InputField.jsx
@@ -1,6 +1,8 @@
+// Дляя використання в формах створених в Formik
+
 import React from 'react';
 import * as S from './InputField.styled';
-import { FormError } from '../FormError';
+import { FormError } from './FormError';
 
 export const InputField = ( { type, name, placeholder, error, isStar, ...props } ) => (
   <>
