+++ src/schemes/modalSchemes/registerScheme.js
